compute: remove the `materialized_view_v2` module
This commit moves the code from the `materialized_view_v2` module into
the `materialized_view` module and removes the former.
